+++ src/FluxFlow.Data/FlowContentCodecMatch.cs
@@ -1,0 +1,8 @@
+namespace FluxFlow.Data;
+
+public enum FlowContentCodecMatch
+{
+    ExactMediaType,
+    StructuredSuffix,
+    MediaFamily
+}