+++ src/FluxFlow.Data/IFlowContentCodec.cs
@@ -1,0 +1,8 @@
+using System.Collections.Immutable;
+
+namespace FluxFlow.Data;
+
+public interface IFlowContentCodec
+{
+    FlowValue Decode(ImmutableArray<byte> content, string? encoding);
+}